Moved existing VTKHDF implementation into a cpp file.
