+++ include/neso_particles/external_interfaces/vtk/vtk.hpp
@@ -153,12 +153,7 @@ public:
   /**
    * Close the HDF5 file. This must be called collectively on the communicator.
    */
-  inline void close() {
-    H5CHK(H5Gclose(this->root));
-    H5CHK(H5Fclose(this->file_id));
-    H5CHK(H5Pclose(this->plist_id));
-    this->is_closed = true;
-  };
+  void close();
 
   /**
    * Create a new VTKHDF file. Must be called collectively on the communicator.
@@ -169,59 +164,7 @@ public:
    * @param dataset_type The type of data the user will write to the file.
    */
   VTKHDF(std::string filename, MPI_Comm comm,
-         std::string dataset_type = "UnstructuredGrid")
-      : comm(comm), is_closed(true), step(0), dataset_type(dataset_type) {
-    MPICHK(MPI_Comm_rank(this->comm, &this->rank));
-    MPICHK(MPI_Comm_size(this->comm, &this->size));
-
-    this->plist_id = H5Pcreate(H5P_FILE_ACCESS);
-    H5CHK(H5Pset_fapl_mpio(this->plist_id, this->comm, MPI_INFO_NULL));
-    H5CHK(this->file_id = H5Fcreate(filename.c_str(), H5F_ACC_TRUNC,
-                                    H5P_DEFAULT, this->plist_id));
-    this->is_closed = false;
-
-    // Create the root group
-    H5CHK(this->root = H5Gcreate(file_id, "VTKHDF", H5P_DEFAULT, H5P_DEFAULT,
-                                 H5P_DEFAULT));
-
-    // Create the global attributes
-    // Version
-    {
-      hid_t property;
-      H5CHK(property = H5Pcreate(H5P_ATTRIBUTE_CREATE));
-      hid_t dataspace;
-      hsize_t dims[1] = {2};
-      H5CHK(dataspace = H5Screate_simple(1, dims, dims));
-      hid_t version;
-      H5CHK(version = H5Acreate(this->root, "Version", H5T_STD_I64LE, dataspace,
-                                property, H5P_DEFAULT));
-      int value[2] = {2, 0};
-      H5CHK(H5Awrite(version, H5T_NATIVE_INT, &value));
-      H5CHK(H5Aclose(version));
-      H5CHK(H5Sclose(dataspace));
-      H5CHK(H5Pclose(property));
-    }
-    // Type
-    {
-      hid_t property;
-      H5CHK(property = H5Pcreate(H5P_ATTRIBUTE_CREATE));
-      hid_t dataspace;
-      H5CHK(dataspace = H5Screate(H5S_SCALAR));
-      hid_t version;
-
-      std::string value = dataset_type;
-      hid_t string_type;
-      H5CHK(string_type = H5Tcreate(H5T_STRING, value.size()));
-      H5CHK(version = H5Acreate(this->root, "Type", string_type, dataspace,
-                                property, H5P_DEFAULT));
-      H5CHK(H5Awrite(version, string_type, value.c_str()));
-
-      H5CHK(H5Tclose(string_type));
-      H5CHK(H5Aclose(version));
-      H5CHK(H5Sclose(dataspace));
-      H5CHK(H5Pclose(property));
-    }
-  }
+         std::string dataset_type = "UnstructuredGrid");
 
   /**
    * Write unstructured data to the file. Must be called collectively on the
@@ -229,131 +172,7 @@ public:
    *
    * @param data Data to write to the file.
    */
-  inline void write(std::vector<UnstructuredCell> &data) {
-    NESOASSERT(this->dataset_type == "UnstructuredGrid",
-               "Attempting to write unstructured grid data to a file which is "
-               "not set up for unstructured grid data.");
-
-    int npoint_local = 0;
-    int ncell_local = 0;
-    std::vector<double> points;
-    std::vector<int> types;
-    types.reserve(data.size());
-    std::vector<int> connectivity;
-    std::vector<int> offsets;
-    std::map<std::string, std::vector<double>> point_data;
-    std::map<std::string, std::vector<double>> cell_data;
-
-    int point_index = 0;
-    offsets.push_back(point_index);
-    for (const auto &ex : data) {
-      const int num_points = ex.num_points;
-      npoint_local += num_points;
-      ncell_local++;
-      NESOASSERT(ex.points.size() == static_cast<size_t>(num_points) * 3,
-                 "Incorrect number of coordinates, expected " +
-                     std::to_string(num_points * 3) + " but found " +
-                     std::to_string(ex.points.size()) + ".");
-
-      points.insert(points.end(), ex.points.begin(), ex.points.end());
-
-      for (auto &name_data : ex.point_data) {
-        auto &n = name_data.first;
-        auto &d = name_data.second;
-        NESOASSERT(d.size() == static_cast<std::size_t>(num_points),
-                   "Points data for " + n + " has wrong array length.");
-        point_data[n].insert(point_data[n].end(), d.begin(), d.end());
-      }
-
-      for (int vx = 0; vx < num_points; vx++) {
-        connectivity.push_back(point_index++);
-      }
-      const int cell_type = static_cast<int>(ex.cell_type);
-      types.push_back(cell_type);
-      offsets.push_back(point_index);
-      for (auto &name_data : ex.cell_data) {
-        cell_data[name_data.first].push_back(name_data.second);
-      }
-    }
-
-    int counts_local[4] = {npoint_local, ncell_local,
-                           static_cast<int>(connectivity.size()),
-                           ncell_local + 1};
-    int offsets_array[4] = {0, 0, 0, 0};
-    int counts_global[4];
-    MPICHK(MPI_Exscan(counts_local, offsets_array, 4, MPI_INT, MPI_SUM,
-                      this->comm));
-    for (int ix = 0; ix < 4; ix++) {
-      counts_global[ix] = offsets_array[ix] + counts_local[ix];
-    }
-    MPICHK(MPI_Bcast(counts_global, 4, MPI_INT, this->size - 1, this->comm));
-
-    // const int nconnectivity_local = counts_local[2];
-    const int noffset_local = counts_local[3];
-
-    const int point_offset = offsets_array[0];
-    const int cell_offset = offsets_array[1];
-    const int connectivity_offset = offsets_array[2];
-    const int offset_offset = offsets_array[3];
-
-    const int npoint_global = counts_global[0];
-    const int ncell_global = counts_global[1];
-    const int nconnectivity_global = counts_global[2];
-    const int noffset_global = counts_global[3];
-
-    NESOASSERT(offsets.size() == static_cast<std::size_t>(noffset_local),
-               "offsets sizes miss-match");
-
-    this->write_dataset_2d(this->root, npoint_global, point_offset,
-                           npoint_local, 3, H5T_IEEE_F64LE, H5T_NATIVE_DOUBLE,
-                           "Points", points);
-
-    std::vector<int> number_of_points = {npoint_local};
-    this->write_dataset(this->root, this->size, this->rank, H5T_STD_I64LE,
-                        H5T_NATIVE_INT, "NumberOfPoints", number_of_points);
-
-    std::vector<int> number_of_cells = {ncell_local};
-    this->write_dataset(this->root, this->size, this->rank, H5T_STD_I64LE,
-                        H5T_NATIVE_INT, "NumberOfCells", number_of_cells);
-
-    this->write_dataset(this->root, ncell_global, cell_offset, H5T_STD_U8LE,
-                        H5T_NATIVE_INT, "Types", types);
-
-    this->write_dataset(this->root, nconnectivity_global, connectivity_offset,
-                        H5T_STD_I64LE, H5T_NATIVE_INT, "Connectivity",
-                        connectivity);
-
-    std::vector<int> number_of_connectivity_ids = {
-        static_cast<int>(connectivity.size())};
-    this->write_dataset(this->root, this->size, this->rank, H5T_STD_I64LE,
-                        H5T_NATIVE_INT, "NumberOfConnectivityIds",
-                        number_of_connectivity_ids);
-
-    this->write_dataset(this->root, noffset_global, offset_offset,
-                        H5T_STD_I64LE, H5T_NATIVE_INT, "Offsets", offsets);
-
-    hid_t point_data_group;
-    H5CHK(point_data_group = H5Gcreate(this->root, "PointData", H5P_DEFAULT,
-                                       H5P_DEFAULT, H5P_DEFAULT));
-    hid_t cell_data_group;
-    H5CHK(cell_data_group = H5Gcreate(this->root, "CellData", H5P_DEFAULT,
-                                      H5P_DEFAULT, H5P_DEFAULT));
-
-    for (auto &name_data : point_data) {
-      this->write_dataset(point_data_group, npoint_global, point_offset,
-                          H5T_IEEE_F64LE, H5T_NATIVE_DOUBLE, name_data.first,
-                          name_data.second);
-    }
-
-    for (auto &name_data : cell_data) {
-      this->write_dataset(cell_data_group, ncell_global, cell_offset,
-                          H5T_IEEE_F64LE, H5T_NATIVE_DOUBLE, name_data.first,
-                          name_data.second);
-    }
-
-    H5CHK(H5Gclose(point_data_group));
-    H5CHK(H5Gclose(cell_data_group));
-  }
+  void write(std::vector<UnstructuredCell> &data);
 };
 
 #else
